:recycle: 非同期処理を改善
